Rag_ChatBot/requirements.txt
python-dotenv
langchain-ollama
langchain-chroma
langchain-community
chromadb
